requirements.txt
django==4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-standardized-errors==0.13.0
pillow==10.3.0
psycopg2-binary==2.9.9
ipython==8.24.0
python-dotenv==1.0.1
pre-commit==3.7.1
black==24.2
coverage==7.5.0
pytest-django==4.7.0
django-cors-headers==4.3.1
flake8==7.0.0
isort==5.13.2
factory-boy==3.3.0
pre-commit==3.7.1